requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==42.0.5
cssbeautifier==1.15.1
cycler==0.12.1
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.8
django-admin-interface==0.28.8
django-allauth==64.1.0
django-allauth-2fa==0.11.1
django-appconf==1.0.6
django-auto-logout==0.5.1
django-browser-reload==1.12.1
django-colorfield==0.11.0
django-compressor==4.5.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-formtools==2.5.1
django-ipware==6.0.5
django-multi-captcha-admin==2.0.0
django-oauth-toolkit==3.0.1
django-otp==1.5.1
django-phonenumber-field==7.3.0
django-qsessions==1.1.5
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-tinymce==4.0.0
django-two-factor-auth==1.16.0
django-use-email-as-username==1.4.0
django-webpack-loader==3.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djlint==1.34.1
djoser==2.2.2
EditorConfig==0.12.4
environ==1.0
filelock==3.15.3
fonttools==4.53.1
frozenlist==1.4.1
geoip2==4.8.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.6
Jinja2==3.1.3
jsbeautifier==1.15.1
json5==0.9.25
jwcrypto==1.5.6
kiwisolver==1.4.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
maxminddb==2.6.2
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.40
phonenumberslite==8.13.40
pillow==10.3.0
pipenv==2024.0.1
platformdirs==4.2.2
prettytable==3.10.0
pycparser==2.22
pygad==3.3.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==2.0.3
python-memcached==1.62
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rcssmin==1.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rjsmin==1.2.2
setuptools==69.2.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.5.0
text-unidecode==1.3
tornado==6.4
tqdm==4.66.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
users==1.0.dev0
virtualenv==20.26.2
wcwidth==0.2.13
whitenoise==6.6.0
yarl==1.9.4
